Refactored range filter and moved things around between columnHelper and FilterHelper
