+++ src/components/AdversariesGearColumnProvider.js
@@ -4,7 +4,6 @@ import {
   makeColumns,
   GENERATED_ID_COL_INDEX,
   indexRender,
-  humanizedNumberRender,
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
