+++ src/components/AdversariesWeaponsColumnProvider.js
@@ -9,6 +9,7 @@ import {
   humanizedNumberRender,
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
+import { getRangeFilterOptions } from "./shared/FilterHelper"
 
 function AdversariesWeaponsColumnProvider({ children, currentBook, metadata }) {
   let bookData = ProvideBookData()
@@ -35,12 +36,14 @@ function AdversariesWeaponsColumnProvider({ children, currentBook, metadata }) {
       options: {
         customBodyRender: (value, tableMeta) =>
           damageRender(value, tableMeta, metadata),
+          ...getRangeFilterOptions("Damage")
       },
     },
     {
       label: "Crit",
       name: "crit",
-      options: { customBodyRender: humanizedNumberRender },
+      options: { customBodyRender: humanizedNumberRender,
+      ...getRangeFilterOptions("Crit") },
     },
     { label: "Range", name: "range" },
     { label: "Special", name: "special" },
