+++ src/components/AttachmentsColumnProvider.js
@@ -2,17 +2,13 @@ import React from "react"
 import Link from "./shared/Link"
 import {
   makeColumns,
-  RESTRICTED_COL_INDEX,
   GENERATED_ID_COL_INDEX,
   indexRender,
   PRICE_FILTER_OPTIONS,
   humanizedNumberRender,
-} from "./shared/ColumnHelper"
-import {
-  getCustomRangeFilterListOptions,
-  getRangeFilterOptions,
   priceRender,
-} from "./shared/FilterHelper"
+} from "./shared/ColumnHelper"
+import { getRangeFilterOptions } from "./shared/FilterHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
 export default function AttachmentsColumnProvider({ children, currentBook }) {
@@ -39,11 +35,7 @@ export default function AttachmentsColumnProvider({ children, currentBook }) {
         label: "Price",
         name: "price",
         options: {
-          customBodyRender: (value, tableMeta) =>
-            `${tableMeta.rowData[RESTRICTED_COL_INDEX] ? "(R) " : ""}${(value &&
-              value.toLocaleString &&
-              value.toLocaleString()) ||
-              value}`,
+          customBodyRender: priceRender,
           ...PRICE_FILTER_OPTIONS,
         },
       },
@@ -51,7 +43,10 @@ export default function AttachmentsColumnProvider({ children, currentBook }) {
       {
         label: "HP",
         name: "hp",
-        options: { customBodyRender: humanizedNumberRender },
+        options: {
+          customBodyRender: humanizedNumberRender,
+          ...getRangeFilterOptions("HP"),
+        },
       },
       { label: "Rarity", name: "rarity", options: { sort: false } },
       { label: "Notes", name: "notes", options: { sort: false } },
