+++ src/components/VehicleAttachmentsColumnProvider.js
@@ -2,16 +2,14 @@ import React from "react"
 import Link from "./shared/Link"
 import {
   makeColumns,
-  RESTRICTED_COL_INDEX,
   GENERATED_ID_COL_INDEX,
   indexRender,
   PRICE_FILTER_OPTIONS,
   humanizedNumberRender,
+  priceRender,
 } from "./shared/ColumnHelper"
 import {
-  getCustomRangeFilterListOptions,
   getRangeFilterOptions,
-  priceRender,
 } from "./shared/FilterHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
@@ -41,23 +39,21 @@ export default function VehicleAttachmentsColumnProvider({
         label: "Price",
         name: "price",
         options: {
-          customBodyRender: (value, tableMeta) =>
-            `${tableMeta.rowData[RESTRICTED_COL_INDEX] ? "(R) " : ""}${(value &&
-              value.toLocaleString &&
-              value.toLocaleString()) ||
-              value}`,
+          customBodyRender: priceRender,
           ...PRICE_FILTER_OPTIONS,
         },
       },
       {
         label: "HP",
         name: "hp",
-        options: { customBodyRender: humanizedNumberRender },
+        options: { customBodyRender: humanizedNumberRender,
+        ...getRangeFilterOptions("HP") },
       },
       {
         label: "Rarity",
         name: "rarity",
-        options: { customBodyRender: humanizedNumberRender },
+        options: { customBodyRender: humanizedNumberRender,
+        ...getRangeFilterOptions("Rarity") },
       },
       {
         label: "Index",
